feat (river_admin): Add state filter to workflow object list
Added 'get_state' function in workflow_view.py to check if there's a state query parameter in the request. Modified the 'list_workflow_objects' function to use this state as a filter when getting a list of workflow objects. Also updated the 'get_objects' method in __init__.py to accept an optional 'state' parameter, and filter the objects if present.