feat (river_admin): Improve state listing by adding slug filter [FMS-2348]
The state_view.py file was updated to improve the search functionality of state listing. Instead of returning all states, the list_states function now uses a keyword ('slug') from the request to filter the states if available. The 'slug' keyword is used to search within slugs of the states. If no slug is provided, it will return all states as before.