feat (river_admin): Add title and description fields to Workflow serializer FMS-2349
Updated the 'CreateWorkflowDto' serializer in the 'river_admin' app to include 'title' and 'description' fields in addition to 'content_type', 'field_name', and 'initial_state'. This change accommodates the inclusion of more descriptive metadata for each workflow. The 'django-river' dependency in requirements.txt was also updated to a newer commit.